test: 외부통신 모듈 fake객체 추가
