+++ src/main/java/kitchenpos/infra/KitchenRiderClientImpl.java
@@ -1,0 +1,14 @@
+package kitchenpos.infra;
+
+import org.springframework.stereotype.Component;
+
+import java.math.BigDecimal;
+import java.util.UUID;
+
+
+@Component
+public class KitchenRiderClientImpl implements KitchenridersClient {
+    @Override
+    public void requestDelivery(UUID orderId, BigDecimal amount, String deliveryAddress) {}
+
+}
