+++ src/main/java/kitchenpos/infra/KitchenridersClient.java
@@ -6,7 +6,7 @@ import java.math.BigDecimal;
 import java.util.UUID;
 
 @Component
-public class KitchenridersClient {
-    public void requestDelivery(final UUID orderId, final BigDecimal amount, final String deliveryAddress) {
-    }
+public interface KitchenridersClient {
+    void requestDelivery(final UUID orderId, final BigDecimal amount, final String deliveryAddress);
+
 }
