+++ src/main/java/kitchenpos/infra/PurgomalumClientImpl.java
@@ -1,0 +1,27 @@
+
+package kitchenpos.infra;
+
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.stereotype.Component;
+import org.springframework.web.client.RestTemplate;
+import org.springframework.web.util.UriComponentsBuilder;
+
+import java.net.URI;
+
+@Component
+public class PurgomalumClientImpl implements PurgomalumClient {
+    private final RestTemplate restTemplate;
+
+    public PurgomalumClientImpl(final RestTemplateBuilder restTemplateBuilder) {
+        this.restTemplate = restTemplateBuilder.build();
+    }
+
+    @Override
+    public boolean containsProfanity(final String text) {
+        final URI url = UriComponentsBuilder.fromUriString("https://www.purgomalum.com/service/containsprofanity")
+            .queryParam("text", text)
+            .build()
+            .toUri();
+        return Boolean.parseBoolean(restTemplate.getForObject(url, String.class));
+    }
+}
