+++ src/test/java/kitchenpos/fake/repository/FakePurogmalumClient.java
@@ -1,0 +1,10 @@
+package kitchenpos.fake.repository;
+
+import kitchenpos.infra.PurgomalumClient;
+
+public class FakePurogmalumClient implements PurgomalumClient {
+    @Override
+    public boolean containsProfanity(String text) {
+        return false;
+    }
+}
